fix: 分辨率往返后栅栏位置/尺寸不再漂移(WM_DISPLAYCHANGE)
游戏全屏会临时改分辨率/DPI,退出后栅栏没回到原始状态。根因是 WM_DISPLAYCHANGE 的破坏性处理:把 w/h 钳进新工作区、x/y clamp 后直接落盘——单向操作,分辨率恢复后无法还原 → 位置永久偏移、互相挤压;且它只更新 f.dpi 却不把 w/h 从 cfg.dpi 换算,DPI 变时几何错配 → 挤压错乱。

修复:WM_DISPLAYCHANGE 只把 w/h 从 cfg.dpi 保逻辑换算到新 DPI(仅钳下限),位置 x/y 保持不动——用户摆放的物理基准不被 clamp 污染,任意分辨率往返严格还原。小分辨率期间部分栅栏处于屏外,恢复后原样回来(全屏游戏本就盖住桌面)。

随提交含 fence 模块拆分重构(行为不变)、版本号对齐 0.1.4。
